fix: split marshal module
