+++ packages/marshal/index.js
@@ -1,14 +1,10 @@
 export {
   REMOTE_STYLE,
   getInterfaceOf,
-  pureCopy,
-  QCLASS,
   getErrorConstructor,
-  sameValueZero,
   passStyleOf,
-  makeMarshal,
-  Remotable,
-  Far,
-} from './src/marshal';
+} from './src/passStyleOf';
+
+export { pureCopy, QCLASS, makeMarshal, Remotable, Far } from './src/marshal';
 
 export { stringify, parse } from './src/marshal-stringify';
